% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{chapter_components}
\alias{chapter_components}
\title{Extract Components of GDPR Chapters/Articles/Points}
\usage{
chapter_components(chapter = NULL, article = NULL, points = NULL)
}
\arguments{
\item{chapter}{chapter from which to extract}

\item{article}{article from which to extract}

\item{points}{points to extract (default all available)}
}
\value{
Either a vector of article titles or a vector of points.
}
\description{
Extract Components of GDPR Chapters/Articles/Points
}
\examples{
chapter_components(chapter = 1) # list all articles in chapter (1)
chapter_components(chapter = 1, article = 1) # extract all points in (1, 1)
chapter_components(chapter = 1, article = 1, points = 2) # extract (1, 1, 2)
}
