% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_data.R
\docType{data}
\name{GDPR_citations}
\alias{GDPR_citations}
\title{GDPR Regulations Citations as a Tidy Object}
\format{A \link{tibble} with 21 rows and 3 variables:
\describe{
\item{number}{Citation number}
\item{text}{Citation text}
\item{type}{"citation"}
}}
\source{
\url{http://openscience.adaptcentre.ie/projects/GDPRtEXT/} Made
available under the
\href{https://creativecommons.org/licenses/by/4.0/}{CC-by-4.0} license
}
\usage{
GDPR_citations
}
\description{
The GDPR regulations citations stored as a \link{tibble}.
}
\keyword{datasets}
