% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_data.R
\docType{data}
\name{GDPR_list}
\alias{GDPR_list}
\title{GDPR Regulations Recitals as a list}
\format{An object of class \code{list} of length 11.}
\source{
\url{http://openscience.adaptcentre.ie/projects/GDPRtEXT/} Made
available under the
\href{https://creativecommons.org/licenses/by/4.0/}{CC-by-4.0} license
}
\usage{
GDPR_list
}
\description{
The GDPR full regulations stored as a list, converted from raw JSON via
\link{fromJSON}.
}
\keyword{datasets}
