% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GDPR.R
\name{get_GDPR}
\alias{get_GDPR}
\title{Download the latest GDPR text}
\source{
\url{http://openscience.adaptcentre.ie/projects/GDPRtEXT/} Made
available under the
\href{https://creativecommons.org/licenses/by/4.0/}{CC-by-4.0} license
}
\usage{
get_GDPR(as = c("list", "data.frame"))
}
\value{
the latest GDPR regulation text as a list. See \code{?}\link{GDPR_data} for a more
detailed description of the returned object.
}
\description{
Use this to obtain a (potentially updated) GDPR regulation text.
}
