% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGDPR-package.R
\docType{package}
\name{tidyGDPR-package}
\alias{tidyGDPR}
\alias{tidyGDPR-package}
\title{tidyGDPR: General Data Protection Regulation as Tidy Object(s)}
\description{
The GDPR recitals, chapters, articles, points, subpoints, and citations as tidy objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonocarroll/tidyGDPR}
  \item Report bugs at \url{https://github.com/jonocarroll/tidyGDPR/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Carroll \email{rpkg@jcarroll.com.au} (0000-0002-1404-5264)

}
\keyword{internal}
