% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_data.R
\docType{data}
\name{GDPR_chapters}
\alias{GDPR_chapters}
\title{GDPR Regulations Chapters as a Tidy Object}
\format{A \link{tibble} with 11 rows and 4 variables:
\describe{
\item{number}{Chapter number (Roman Numerals)}
\item{title}{Chapter title}
\item{type}{"chapter"}
\item{contents}{A \link{data.frame} of additional nested contents,
from articles down to subpoints}
}

The \code{contents} column \link{data.frame} is a nested object, itself containing a
\code{number}, \code{title}, \code{type}, and \code{contents} column, corresponding to article
numbers. That \code{contents} column contains further nested \link{data.frame}s of
points and subpoints, which have not been de-nested.

See \link{chapter_components} for an accessible method.}
\source{
\url{http://openscience.adaptcentre.ie/projects/GDPRtEXT/} Made
available under the
\href{https://creativecommons.org/licenses/by/4.0/}{CC-by-4.0} license
}
\usage{
GDPR_chapters
}
\description{
The GDPR regulations chapters stored as a (nested) \link{tibble}.
}
\keyword{datasets}
